requirements.txt
absl-py==0.13.0
appnope==0.1.2
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
batch-face==1.0.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cloudpickle==1.6.0
cycler==0.10.0
decorator==5.0.9
eli5==0.11.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.16
grpcio==1.34.1
h5py==3.1.0
idna==2.10
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keras-tuner==1.0.3
keras-vggface==0.6
kiwisolver==1.3.1
kt-legacy==1.0.3
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
numba==0.53.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.2.54
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
shap==0.39.0
six==1.15.0
slicer==0.0.7
tabulate==0.8.9
temp==2020.7.2
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
